Simplify app structure, remove ejs
