+++ examples/backend-access-request/src/utils/getConfig.ts
