+++ examples/backend-access-request/src/views/escapeHtml.ts
@@ -1,0 +1,10 @@
+export function escapeHtml(text: string) {
+    let lookup: Record<string, string> = {
+        '&': "&amp;",
+        '"': "&quot;",
+        '\'': "&apos;",
+        '<': "&lt;",
+        '>': "&gt;"
+    };
+    return text.replace( /[&"'<>]/g, c => lookup[c] );
+}
