--- examples/backend-access-request/views/granted.ejs
@@ -1,4 +1,0 @@
-<h2>Access Grant</h2>
-<pre><code><%= JSON.stringify(accessGrant, undefined, "  ") %></code></pre>
-<h2>Resource content</h2>
-<pre><code><%= fileContent %></code></pre>
